Simplify the implementation by removing the transaction builder inherited class... It felt wrong from the start
