+++ src/Stratis.Bitcoin.Features.SidechainWallet/TransactionBuildContext.cs
@@ -36,6 +36,5 @@ namespace Stratis.Bitcoin.Features.SidechainWallet
         }
 
         public string SidechainIdentifier { get; set; }
-        public new SidechainWallet.TransactionBuilder TransactionBuilder { get; set; }
     }
 }
